+++ SpatialGDK/Source/SpatialGDKEditorToolbar/Private/SpatialGDKEditorToolbar.cpp
@@ -26,7 +26,6 @@
 #include "HAL/FileManager.h"
 #include "Sound/SoundBase.h"
 
-#include "Algo/Accumulate.h"
 #include "AssetRegistryModule.h"
 #include "GeneralProjectSettings.h"
 #include "LevelEditor.h"
@@ -364,9 +363,7 @@ bool FSpatialGDKEditorToolbarModule::ValidateGeneratedLaunchConfig() const
 	}
 
 	const ULevelEditorPlaySettings* LevelEditorPlaySettings = GetDefault<ULevelEditorPlaySettings>();
-	const int32 ServerCount = Algo::Accumulate(LevelEditorPlaySettings->WorkerTypesToLaunch, 0, [](int32 Current, const TPair<FName, int32>& Element) { return Current + Element.Value; });
-	
-	if (!SpatialGDKRuntimeSettings->bEnableHandover && ServerCount > 1)
+	if (!SpatialGDKRuntimeSettings->bEnableHandover && LevelEditorPlaySettings->GetTotalServerWorkerCount() > 1)
 	{
 		const EAppReturnType::Type Result = FMessageDialog::Open(EAppMsgType::YesNo, FText::FromString(TEXT("Property handover is disabled and multiple launch servers are specified.\nThis is not supported.\n\nDo you want to configure your project settings now?")));